requirement.txt
python-dotenv
supabase
requests
uvicorn
fastapi